requirements.txt
faiss-cpu
Flask
requests
langchain-community
numpy
openai
pytest
python-dotenv
tiktoken
